% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoDB.R
\name{pseudoDB}
\alias{pseudoDB}
\title{Pseudo-database class}
\description{
Pseudo-database class

Pseudo-database class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pseudoDB$new()}}
\item \href{#method-read_config}{\code{pseudoDB$read_config()}}
\item \href{#method-check_files_exist}{\code{pseudoDB$check_files_exist()}}
\item \href{#method-open_sqlite}{\code{pseudoDB$open_sqlite()}}
\item \href{#method-log}{\code{pseudoDB$log()}}
\item \href{#method-read_data}{\code{pseudoDB$read_data()}}
\item \href{#method-read_data_fread}{\code{pseudoDB$read_data_fread()}}
\item \href{#method-clone}{\code{pseudoDB$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoDB$new(config_file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_config"></a>}}
\if{latex}{\out{\hypertarget{method-read_config}{}}}
\subsection{Method \code{read_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoDB$read_config(fn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_files_exist"></a>}}
\if{latex}{\out{\hypertarget{method-check_files_exist}{}}}
\subsection{Method \code{check_files_exist()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoDB$check_files_exist()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-open_sqlite"></a>}}
\if{latex}{\out{\hypertarget{method-open_sqlite}{}}}
\subsection{Method \code{open_sqlite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoDB$open_sqlite()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log"></a>}}
\if{latex}{\out{\hypertarget{method-log}{}}}
\subsection{Method \code{log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoDB$log(what)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_data"></a>}}
\if{latex}{\out{\hypertarget{method-read_data}{}}}
\subsection{Method \code{read_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoDB$read_data(fn)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fn}}{Bare filename to read (full path is read from config)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_data_fread"></a>}}
\if{latex}{\out{\hypertarget{method-read_data_fread}{}}}
\subsection{Method \code{read_data_fread()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoDB$read_data_fread(fn, quote, sep, fill, skip, encoding)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pseudoDB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
